move pqBuildScoreProvider() into BSP
